recommendation-service/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
sentence-transformers
scikit-learn
numpy
python-dotenv
pydantic
